+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/document/bulk/BulkRequestProtoUtils.java
@@ -90,9 +90,9 @@ public class BulkRequestProtoUtils {
             throw new UnsupportedOperationException("type param is not supported");
         }
 
-        // TODO support global_params
-        if (request.hasGlobalParams()) {
-            throw new UnsupportedOperationException("global_params param is not supported yet");
+        // TODO support global_params.human parameter
+        if (request.getGlobalParams().hasHuman()) {
+            throw new UnsupportedOperationException("global_params.human param is not supported yet");
         }
 
         return bulkRequest;