+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/search/SearchRequestProtoUtils.java
@@ -189,9 +189,9 @@ public class SearchRequestProtoUtils {
             throw new UnsupportedOperationException("typed_keys param is not supported yet");
         }
 
-        // TODO support global_params
-        if (request.hasGlobalParams()) {
-            throw new UnsupportedOperationException("global_params param is not supported yet");
+        // TODO support global_params.human parameter
+        if (request.getGlobalParams().hasHuman()) {
+            throw new UnsupportedOperationException("global_params.human param is not supported yet");
         }
     }