+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/response/exceptions/shardoperationfailedexception/ReplicationResponseShardInfoFailureProtoUtils.java
@@ -10,6 +10,7 @@ package org.opensearch.transport.grpc.proto.response.exceptions.shardoperationfa
 import org.opensearch.action.support.replication.ReplicationResponse;
 import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.XContentBuilder;
+import org.opensearch.protobufs.GlobalParams;
 import org.opensearch.protobufs.ShardFailure;
 import org.opensearch.transport.grpc.proto.response.exceptions.opensearchexception.OpenSearchExceptionProtoUtils;
 
@@ -29,9 +30,10 @@ public class ReplicationResponseShardInfoFailureProtoUtils {
      * This method is overridden by various exception classes, which are hardcoded here.
      *
      * @param exception The ReplicationResponse.ShardInfo.Failure to convert metadata from
+     * @param params The global gRPC request parameters
      * @return A map containing the exception's metadata as ObjectMap.Value objects
      */
-    public static ShardFailure toProto(ReplicationResponse.ShardInfo.Failure exception) throws IOException {
+    public static ShardFailure toProto(ReplicationResponse.ShardInfo.Failure exception, GlobalParams params) throws IOException {
         ShardFailure.Builder shardFailure = ShardFailure.newBuilder();
         if (exception.index() != null) {
             shardFailure.setIndex(exception.index());
@@ -40,7 +42,7 @@ public class ReplicationResponseShardInfoFailureProtoUtils {
         if (exception.nodeId() != null) {
             shardFailure.setNode(exception.nodeId());
         }
-        shardFailure.setReason(OpenSearchExceptionProtoUtils.generateThrowableProto(exception.getCause()));
+        shardFailure.setReason(OpenSearchExceptionProtoUtils.generateThrowableProto(exception.getCause(), params));
         shardFailure.setStatus(exception.status().name());
         shardFailure.setPrimary(exception.primary());
         return shardFailure.build();