+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/services/SearchServiceImpl.java
@@ -16,6 +16,7 @@ import org.opensearch.transport.grpc.listeners.SearchRequestActionListener;
 import org.opensearch.transport.grpc.proto.request.search.SearchRequestProtoUtils;
 import org.opensearch.transport.grpc.proto.request.search.query.AbstractQueryBuilderProtoUtils;
 import org.opensearch.transport.grpc.util.GrpcErrorHandler;
+import org.opensearch.transport.grpc.util.GrpcParamsHandler;
 
 import java.io.IOException;
 
@@ -61,14 +62,14 @@ public class SearchServiceImpl extends SearchServiceGrpc.SearchServiceImplBase {
         org.opensearch.protobufs.SearchRequest request,
         StreamObserver<org.opensearch.protobufs.SearchResponse> responseObserver
     ) {
-
         try {
+            GrpcParamsHandler.validateStackTraceDetailsConfiguration(request.getGlobalParams());
             org.opensearch.action.search.SearchRequest searchRequest = SearchRequestProtoUtils.prepareRequest(request, client, queryUtils);
-            SearchRequestActionListener listener = new SearchRequestActionListener(responseObserver);
+            SearchRequestActionListener listener = new SearchRequestActionListener(responseObserver, request.getGlobalParams());
             client.search(searchRequest, listener);
         } catch (RuntimeException | IOException e) {
             logger.debug("SearchServiceImpl failed to process search request, request=" + request + ", error=" + e.getMessage());
-            StatusRuntimeException grpcError = GrpcErrorHandler.convertToGrpcError(e);
+            StatusRuntimeException grpcError = GrpcErrorHandler.convertToGrpcError(e, request.getGlobalParams());
             responseObserver.onError(grpcError);
         }
     }